+++ src/test/java/fr/marcwrobel/jbanking/creditor/CreditorIdentifierTest.java
@@ -34,24 +34,24 @@ class CreditorIdentifierTest {
       "SM94ZZZA1B2C3D4E5F6G7H8", "ES50ZZZM23456789", "LU27ZZZ0000000000123456789", "CH1312300000012345",
       "MT31ZZZ123456789X", "EE49ZZZEE00012345678");
 
-  private static final String VALID_CI_COUNTRY = "FR";
+  private static final IsoCountry VALID_CI_COUNTRY = IsoCountry.FR;
   private static final String VALID_CI_CHECKDIGIT = "72";
   private static final String VALID_CI_BUSINESS_CODE = "ZZZ";
   private static final String VALID_CI_NATIONAL_ID = "123456";
-  private static final String VALID_CI = VALID_CI_COUNTRY + VALID_CI_CHECKDIGIT + VALID_CI_BUSINESS_CODE
+  private static final String VALID_CI = VALID_CI_COUNTRY.getAlpha2Code() + VALID_CI_CHECKDIGIT + VALID_CI_BUSINESS_CODE
       + VALID_CI_NATIONAL_ID;
   private static final String VALID_CI2 = "BE69ZZZ050D000000008";
 
   private static final String INVALID_CI_NATIONAL_ID = "132!";
 
-  private static final String CI_WITH_INVALID_FORMAT = VALID_CI_COUNTRY + VALID_CI_CHECKDIGIT + VALID_CI_BUSINESS_CODE
-      + INVALID_CI_NATIONAL_ID;
+  private static final String CI_WITH_INVALID_FORMAT = VALID_CI_COUNTRY.getAlpha2Code() + VALID_CI_CHECKDIGIT
+      + VALID_CI_BUSINESS_CODE + INVALID_CI_NATIONAL_ID;
   private static final String CI_WITH_UNKNOWN_COUNTRY = "ZZ" + VALID_CI_CHECKDIGIT + VALID_CI_BUSINESS_CODE
       + VALID_CI_NATIONAL_ID;
   private static final String CI_WITH_UNSUPPORTED_COUNTRY = "US" + VALID_CI_CHECKDIGIT + VALID_CI_BUSINESS_CODE
       + VALID_CI_NATIONAL_ID;
 
-  private static final String CI_WITH_INVALID_CHECK_DIGIT = VALID_CI_COUNTRY + "09" + VALID_CI_BUSINESS_CODE
+  private static final String CI_WITH_INVALID_CHECK_DIGIT = VALID_CI_COUNTRY.getAlpha2Code() + "09" + VALID_CI_BUSINESS_CODE
       + VALID_CI_NATIONAL_ID;
 
   @Test
@@ -162,7 +162,8 @@ class CreditorIdentifierTest {
   void validCreditorIdentifierDecomposition() {
     assertTrue(CreditorIdentifier.isValid(VALID_CI));
     CreditorIdentifier creditorId = new CreditorIdentifier(VALID_CI);
-    assertEquals(VALID_CI_COUNTRY, creditorId.getCountryCode());
+    assertEquals(VALID_CI_COUNTRY.getAlpha2Code(), creditorId.getCountryCode());
+    assertEquals(VALID_CI_COUNTRY, creditorId.getCountry());
     assertEquals(VALID_CI_CHECKDIGIT, creditorId.getCheckDigit());
     assertEquals(VALID_CI_BUSINESS_CODE, creditorId.getBusinessCode());
     assertEquals(VALID_CI_NATIONAL_ID, creditorId.getNationalIdentifier());